questme/website/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.3
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
filelock==3.16.1
fonttools==4.55.3
html5lib==1.1
idna==3.10
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.0
numpy==2.2.1
opencv-python==4.11.0.86
oscrypto==1.3.0
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
pygame==2.5.2
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyMuPDF==1.25.5
pyparsing==3.2.1
pypdf==5.4.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
virtualenv==20.28.1
webencodings==0.5.1
xhtml2pdf==0.2.17
